refactored processOptions and runModel into run.ts
